• feat(chatgpt): use inquirer for interactive prompt • fix(bash): can be used from everywhere
