+++ index.js
@@ -1,15 +1,11 @@
-import { Configuration, OpenAIApi } from "@fortaine/openai";
 import { parseArgs } from "node:util";
-import { streamCompletion } from "@fortaine/openai/stream";
+import process from "node:process";
+import { fileURLToPath } from "node:url";
 
+import inquirer from "inquirer";
 import dotenv from "dotenv-flow";
-import process from "process";
 
-import path from "path";
-import { fileURLToPath } from "url";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+const __dirname = fileURLToPath(new URL(".", import.meta.url));
 
 process.chdir(__dirname);
 
@@ -18,70 +14,34 @@ dotenv.config({
   silent: true,
 });
 
-const configuration = new Configuration({
-  apiKey: process.env.OPENAI_API_KEY,
-});
-const openai = new OpenAIApi(configuration);
-
 // https://2ality.com/2022/08/node-util-parseargs.html
 const {
-  values: { prompt },
+  values: { model },
 } = parseArgs({
   options: {
-    prompt: {
+    model: {
       type: "string",
     },
   },
 });
 
-if (typeof prompt === "undefined") {
-  throw new TypeError(
-    "[ERR_PARSE_ARGS_INVALID_OPTION_VALUE]: Option '--prompt <value>' argument missing"
-  );
-}
-
-try {
-  // https://help.openai.com/en/articles/4936856-what-are-tokens-and-how-to-count-them
-  const max_tokens = 4097 - prompt.length;
-
-  const completion = await openai.createCompletion(
-    {
-      model: "text-davinci-003",
-      max_tokens,
-      prompt,
-      stream: true,
-    },
-    { responseType: "stream" }
-  );
-
-  for await (const message of streamCompletion(completion.data)) {
-    try {
-      const parsed = JSON.parse(message);
-      const { text } = parsed.choices[0];
-
-      process.stdout.write(text);
-    } catch (error) {
-      console.error("Could not JSON parse stream message", message, error);
-    }
-  }
-
-  process.stdout.write("\n");
-} catch (error) {
-  if (error.response?.status) {
-    console.error(error.response.status, error.message);
-
-    for await (const data of error.response.data) {
-      const message = data.toString();
-
-      try {
-        const parsed = JSON.parse(message);
-
-        console.error("An error occurred during OpenAI request: ", parsed);
-      } catch (error) {
-        console.error("An error occurred during OpenAI request: ", message);
+const questions = [
+  {
+    type: "input",
+    name: "prompt",
+    message: "prompt",
+    validate(text) {
+      if (text.length === 0) {
+        return "[ERR_PARSE_ARGS_INVALID_OPTION_VALUE]: Option '--prompt <value>' argument missing";
       }
-    }
-  } else {
-    console.error("An error occurred during OpenAI request", error);
-  }
-}
+
+      return true;
+    },
+  },
+];
+
+const { prompt } = await inquirer.prompt(questions);
+
+const { createCompletion } = await import("./createCompletion.js");
+
+createCompletion({ model, prompt });
